requirements.txt
praw>=7.7.0
python-dotenv>=0.19.0
